Revamp of pyside MoSpaces object
